cnn/pytorch/requirements.txt
# Pillow==10.4.0
Pillow==9.5 # AttributeError: 'FreeTypeFont' object has no attribute 'getsize'
pandas==2.0.3
torch==2.4.0
torchvision==0.19.0
ipython==8.12.3
torchaudio==2.4.0
matplotlib==3.7.5
scikit-learn==1.3.2